// src/models/apflora/Ap.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AeTaxonomiesId } from './AeTaxonomies';
import type { AdresseId } from './Adresse';
import type { ProjektId } from './Projekt';

/** Identifier type for apflora.ap */
export type ApId = string & { __brand: 'apflora.ap' };

/**
 * Represents the table apflora.ap
 * Hier wird bestimmt, welche Benutzer mit den rollen "apflora_ap_writer" oder "apflora_reader" Zugriff auf einen AP erhalten
 */
export default interface Ap {
  /** Primärschlüssel */
  id: ApId;

  /** In welchem Bearbeitungsstand befindet sich der AP? */
  bearbeitung: number | null;

  /** Wann wurde mit der Umsetzung des Aktionsplans begonnen? */
  start_jahr: number | null;

  /** In welchem Umsetzungsstand befindet sich der AP? */
  umsetzung: number | null;

  /** Von wem wurde der Datensatz zuletzt geändert? */
  changed_by: string | null;

  /** Namensgebende Art. Unter ihrem Namen bzw. Nummer werden Kontrollen an InfoFlora geliefert */
  art_id: AeTaxonomiesId | null;

  /** Zugehöriger Bearbeiter. Fremdschlüssel aus der Tabelle "adresse" */
  bearbeiter: AdresseId | null;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "projekt" */
  proj_id: ProjektId | null;

  /** bester Beobachtungszeitpunkt */
  ekf_beobachtungszeitpunkt: string | null;

  created_at: Date;

  updated_at: Date;
}

/**
 * Represents the initializer for the table apflora.ap
 * Hier wird bestimmt, welche Benutzer mit den rollen "apflora_ap_writer" oder "apflora_reader" Zugriff auf einen AP erhalten
 */
export interface ApInitializer {
  /**
   * Primärschlüssel
   * Default value: uuid_generate_v1mc()
   */
  id?: ApId;

  /** In welchem Bearbeitungsstand befindet sich der AP? */
  bearbeitung?: number | null;

  /** Wann wurde mit der Umsetzung des Aktionsplans begonnen? */
  start_jahr?: number | null;

  /** In welchem Umsetzungsstand befindet sich der AP? */
  umsetzung?: number | null;

  /**
   * Von wem wurde der Datensatz zuletzt geändert?
   * Default value: NULL::character varying
   */
  changed_by?: string | null;

  /** Namensgebende Art. Unter ihrem Namen bzw. Nummer werden Kontrollen an InfoFlora geliefert */
  art_id?: AeTaxonomiesId | null;

  /** Zugehöriger Bearbeiter. Fremdschlüssel aus der Tabelle "adresse" */
  bearbeiter?: AdresseId | null;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "projekt" */
  proj_id?: ProjektId | null;

  /** bester Beobachtungszeitpunkt */
  ekf_beobachtungszeitpunkt?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;
}

/**
 * Represents the mutator for the table apflora.ap
 * Hier wird bestimmt, welche Benutzer mit den rollen "apflora_ap_writer" oder "apflora_reader" Zugriff auf einen AP erhalten
 */
export interface ApMutator {
  /** Primärschlüssel */
  id?: ApId;

  /** In welchem Bearbeitungsstand befindet sich der AP? */
  bearbeitung?: number | null;

  /** Wann wurde mit der Umsetzung des Aktionsplans begonnen? */
  start_jahr?: number | null;

  /** In welchem Umsetzungsstand befindet sich der AP? */
  umsetzung?: number | null;

  /** Von wem wurde der Datensatz zuletzt geändert? */
  changed_by?: string | null;

  /** Namensgebende Art. Unter ihrem Namen bzw. Nummer werden Kontrollen an InfoFlora geliefert */
  art_id?: AeTaxonomiesId | null;

  /** Zugehöriger Bearbeiter. Fremdschlüssel aus der Tabelle "adresse" */
  bearbeiter?: AdresseId | null;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "projekt" */
  proj_id?: ProjektId | null;

  /** bester Beobachtungszeitpunkt */
  ekf_beobachtungszeitpunkt?: string | null;

  created_at?: Date;

  updated_at?: Date;
}